+++ src/main/java/com/companystrator/order/dto/req/PlaceOrderItemDTO.java
@@ -10,7 +10,7 @@ public record PlaceOrderItemDTO(
     @JsonProperty("product_code")
     @NotNull(message = "is mandatory")
     @Digits(integer = 10, fraction = 0, message = "Code must be exactly 10 digits")
-    int productCode,
+    Long productCode,
 
     @JsonProperty("quantity")
     @Positive(message = "Product amount USD must be greater than zero")
